Still more changes on UnitSelection code
git-svn-id: https://mary.opendfki.de/repos/trunk@93 953a6561-930b-0410-b2a6-db37d1b2ae63

+++ java/de/dfki/lt/mary/unitselection/TargetCostFunction.java
@@ -48,7 +48,7 @@ public interface TargetCostFunction
      * @param unit
      * @return a non-negative number; smaller values mean better fit, i.e. smaller cost.
      */
-    public int cost(Target target, Unit unit);
+    public float cost(Target target, Unit unit);
     
     /**
      * Set the features of the cost function
